reorganize managed keys for better error messaging
